move interpreter-based SDK version resolution to pkg/agentfs
Pure relocation, no behavior change: the Python/Node binary finders and
installed-version probes (plus the embedded node_resolve_version.js)
move from cmd/lk into pkg/agentfs as exported functions, with
agentfs.ResolveInstalledSDKVersion as the shared entry point, so
callers outside the startup checks can reuse the resolver.
