+++ pkg/agentfs/node_resolve_version.js
